feat: adds phpunit `12.2.1` support
