+++ composer.json
@@ -20,16 +20,15 @@
         "php": "^8.3.0",
         "brianium/paratest": "^7.10.1",
         "nunomaduro/collision": "^8.8.0",
-        "nunomaduro/pokio": "dev-main",
         "nunomaduro/termwind": "^2.3.1",
         "pestphp/pest-plugin": "^4.0.0",
         "pestphp/pest-plugin-arch": "^4.0.0",
         "pestphp/pest-plugin-mutate": "^4.0.0",
-        "phpunit/phpunit": "^12.1.6"
+        "phpunit/phpunit": "^12.2.1"
     },
     "conflict": {
         "filp/whoops": "<2.16.0",
-        "phpunit/phpunit": ">12.1.6",
+        "phpunit/phpunit": ">12.2.1",
         "sebastian/exporter": "<7.0.0",
         "webmozart/assert": "<1.11.0"
     },
