+++ tests/Features/Covers/ExceptionHandling.php
@@ -1,0 +1,10 @@
+<?php
+
+use Pest\PendingCalls\TestCall;
+use Pest\TestSuite;
+
+it('throws exception if no class nor method has been found', function () {
+    $testCall = new TestCall(TestSuite::getInstance(), 'filename', 'description', fn () => 'closure');
+
+    $testCall->covers('fakeName');
+})->throws(InvalidArgumentException::class, 'No class, trait or method named "fakeName" has been found.');
