+++ tests/Features/Covers/FunctionCoverage.php
@@ -1,0 +1,12 @@
+<?php
+
+use PHPUnit\Framework\Attributes\CoversFunction;
+
+function testCoversFunction() {}
+
+it('uses the correct PHPUnit attribute for function', function () {
+    $attributes = (new ReflectionClass($this))->getAttributes();
+
+    expect($attributes[1]->getName())->toBe(CoversFunction::class);
+    expect($attributes[1]->getArguments()[0])->toBe('testCoversFunction');
+})->coversFunction('testCoversFunction');
